fix(solves): int-cast warnings, dead code, missing return

+++ solves/acmp/simple_math/10.cpp
@@ -13,6 +13,6 @@ signed main() {
     }
 
     for (int i = 0; i < (int) ans.size(); i++) {
-        (i == ans.size() - 1) ? cout << ans[i] : cout << ans[i] << ' ';
+        (i == (int) ans.size() - 1) ? cout << ans[i] : cout << ans[i] << ' ';
     }
 }
